+++ apps/constants.py
@@ -75,5 +75,4 @@ THEME_CHOICES = (
 # 判断节点在线时间间隔
 NODE_TIME_OUT = 75
 
-
 DEFAULT_CACHE_TTL = 60 * 60 * 2
