+++ apps/sspanel/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from . import views
+from apps.sspanel import views, admin_views
 
 
 app_name = "sspanel"
@@ -34,40 +34,53 @@ urlpatterns = [
     # 推广相关
     path("aff/invite/", views.AffInviteView.as_view(), name="aff_invite"),
     path("aff/status/", views.AffStatusView.as_view(), name="aff_status"),
-    # 网站后台面板
-    path("backend/", views.system_status, name="system_status"),
+    # ====================================================================
+    # 网站后台界面
+    # ====================================================================
+    path(
+        "backend/user_online_ip_log/",
+        admin_views.UserOnlineIpLogView.as_view(),
+        name="user_online_ip_log",
+    ),
+    path("backend/", admin_views.system_status, name="system_status"),
     # 邀请码相关
-    path("backend/invite/", views.backend_invite, name="backend_invite"),
-    path("invite_gen_code/", views.gen_invite_code, name="geninvitecode"),
+    path("backend/invite/", admin_views.backend_invite, name="backend_invite"),
+    path("invite_gen_code/", admin_views.gen_invite_code, name="geninvitecode"),
     # 节点相关
-    path("backend/nodeinfo/", views.backend_node_info, name="backend_node_info"),
-    path("backend/node/delete/<int:node_id>/", views.node_delete, name="node_delete"),
-    path("backend/node/edit/<int:node_id>/", views.node_edit, name="node_edit"),
-    path("backend/node/create/", views.node_create, name="node_create"),
+    path("backend/nodeinfo/", admin_views.backend_node_info, name="backend_node_info"),
+    path(
+        "backend/node/delete/<int:node_id>/",
+        admin_views.node_delete,
+        name="node_delete",
+    ),
+    path("backend/node/edit/<int:node_id>/", admin_views.node_edit, name="node_edit"),
+    path("backend/node/create/", admin_views.node_create, name="node_create"),
     # 用户相关
-    path("backend/userlist/", views.backend_userlist, name="user_list"),
-    path("backend/user/delete/<int:pk>/", views.user_delete, name="user_delete"),
-    path("backend/user/search/", views.user_search, name="user_search"),
-    path("backend/user/status/", views.user_status, name="user_status"),
+    path("backend/userlist/", admin_views.backend_userlist, name="user_list"),
+    path("backend/user/delete/<int:pk>/", admin_views.user_delete, name="user_delete"),
+    path("backend/user/search/", admin_views.user_search, name="user_search"),
+    path("backend/user/status/", admin_views.user_status, name="user_status"),
     # 商品充值相关
-    path("backend/charge/", views.backend_charge, name="backend_charge"),
-    path("backend/shop/", views.backend_shop, name="backend_shop"),
-    path("backend/shop/delete/<int:pk>/", views.good_delete, name="good_delete"),
-    path("backend/good/create/", views.good_create, name="good_create"),
-    path("backend/good/edit/<int:pk>/", views.good_edit, name="good_edit"),
-    path("backend/purchase/history/", views.purchase_history, name="purchase_history"),
+    path("backend/charge/", admin_views.backend_charge, name="backend_charge"),
+    path("backend/shop/", admin_views.backend_shop, name="backend_shop"),
+    path("backend/shop/delete/<int:pk>/", admin_views.good_delete, name="good_delete"),
+    path("backend/good/create/", admin_views.good_create, name="good_create"),
+    path("backend/good/edit/<int:pk>/", admin_views.good_edit, name="good_edit"),
+    path(
+        "backend/purchase/history/",
+        admin_views.purchase_history,
+        name="purchase_history",
+    ),
     # 公告管理相关
-    path("backend/anno/", views.backend_anno, name="backend_anno"),
-    path("backend/anno/delete/<int:pk>/", views.anno_delete, name="anno_delete"),
-    path("backend/anno/create/", views.anno_create, name="anno_create"),
-    path("backend/anno/edit/<int:pk>/", views.anno_edit, name="anno_edit"),
+    path("backend/anno/", admin_views.backend_anno, name="backend_anno"),
+    path("backend/anno/delete/<int:pk>/", admin_views.anno_delete, name="anno_delete"),
+    path("backend/anno/create/", admin_views.anno_create, name="anno_create"),
+    path("backend/anno/edit/<int:pk>/", admin_views.anno_edit, name="anno_edit"),
     # 工单相关
-    path("backend/ticket/", views.backend_ticket, name="backend_ticket"),
+    path("backend/ticket/", admin_views.backend_ticket, name="backend_ticket"),
     path(
         "backend/ticket/edit/<int:pk>/",
-        views.backend_ticketedit,
+        admin_views.backend_ticketedit,
         name="backend_ticketedit",
     ),
-    # 在线ip
-    path("backend/aliveuser/", views.backend_alive_user, name="alive_user"),
 ]
