+++ apps/ssserver/admin.py
@@ -30,14 +30,8 @@ class NodeOnlineAdmin(admin.ModelAdmin):
     list_display = ["node_id", "online_user"]
 
 
-class AliveIpAdmin(admin.ModelAdmin):
-    list_display = ["node_id", "user", "ip", "log_time"]
-    list_filter = ["node_id", "log_time"]
-
-
 # Register your models here.
 admin.site.register(models.Node, NodeAdmin)
 admin.site.register(models.Suser, SUserAdmin)
-admin.site.register(models.AliveIp, AliveIpAdmin)
 admin.site.register(models.TrafficLog, TrafficLogAdmin)
 admin.site.register(models.NodeOnlineLog, NodeOnlineAdmin)
