+++ templates/backendbase.html
@@ -60,7 +60,7 @@
                     <a class="navbar-item" href="{% url 'sspanel:backend_ticket' %}">工单管理</a>
                     <a class="navbar-item is-active">节点管理</a>
                     <a class="navbar-item" href="{% url 'sspanel:backend_node_info' %}">节点列表</a>
-                    <a class="navbar-item" href="{% url 'sspanel:alive_user' %}">在线IP</a>
+                    <a class="navbar-item" href="{% url 'sspanel:user_online_ip_log' %}">在线IP</a>
                     <a class="navbar-item is-active">交易管理</a>
                     <a class="navbar-item" href="{% url 'sspanel:backend_charge' %}">充值管理</a>
                     <a class="navbar-item" href="{% url 'sspanel:backend_shop' %}">商品管理</a>
@@ -122,7 +122,7 @@
                                 <a href="{% url 'sspanel:backend_node_info' %}">节点列表</a>
                             </li>
                             <li>
-                                <a href="{% url 'sspanel:alive_user' %}">在线IP</a>
+                                <a href="{% url 'sspanel:user_online_ip_log' %}">在线IP</a>
                             </li>
                         </ul>
                         <p class="menu-label">
