+++ v2/__init__.py
@@ -1,1 +1,1 @@
-default_app_config = 'v2.apps.V2Config'
+default_app_config = "v2.apps.V2Config"
